Time: 0 ms (100.00%) | Memory: 45.3 MB (22.30%) - LeetSync

+++ 98-validate-binary-search-tree/validate-binary-search-tree.java
@@ -19,11 +19,11 @@ class Solution {
         //if(root.val == 2147483647 || root.val == -2147483647) return true;
         return validateBST(root, Long.MIN_VALUE, Long.MAX_VALUE);
     }
-    private boolean validateBST(TreeNode root,Long low,Long high){
+    private boolean validateBST(TreeNode root,long low,long high){
         if(root==null) return true;
         if(root.val<=low || root.val>=high) return false;
-        boolean left = validateBST(root.left, low,(long) root.val);
-        boolean right = validateBST(root.right, (long)root.val, high);
+        boolean left = validateBST(root.left, low, root.val);
+        boolean right = validateBST(root.right, root.val, high);
         return left&&right;
 
     }
